container_app/requirements.txt
# Terprint Menu Downloader Container App Requirements
# Note: terprint-menu-downloader is installed from local source in Dockerfile

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Scheduling
apscheduler>=3.10.4

# Azure SDK dependencies
azure-identity>=1.15.0
azure-storage-file-datalake>=12.14.0
azure-storage-blob>=12.19.0

# HTTP requests
requests>=2.31.0
httpx>=0.26.0

# Environment variables
python-dotenv>=1.0.0

# OpenTelemetry for Azure Monitor
opentelemetry-instrumentation-fastapi>=0.43b0
azure-monitor-opentelemetry>=1.2.0
